# Use a separate thread for tiered compilation background work